pyproject.toml
[project]
name = "stock-prediction-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.39",
    "duckdb>=1.5.0",
    "feedparser>=6.0.12",
    "loguru>=0.7.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "pytorch-lightning>=2.6.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "shap>=0.51.0",
    "streamlit>=1.19.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "xgboost>=3.2.0",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.setuptools.packages.find]
where = ["src"]
